Feat: 포스팅 수정과 삭제시 이미지데이터 변경안되는 문제 수정

+++ src/main/java/com/project/blog/domain/user/service/UserService.java
@@ -159,7 +159,10 @@ public class UserService {
             throw new CustomException(ExceptionType.PASSWORD_NOT_CORRECT);
         }
 
-        imageService.deleteImageFromS3(user.getImageId());
+        if (user.getProfileImage() != null) {
+            imageService.deleteImageFromS3(user.getImageId());
+        }
+
         userRepository.delete(user);
     }
